Split book into multiple components
