+++ src/components/Progress.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function Progress() {
+  return (
+    <div className="progress">
+      <h5>CURRENT CHAPTER</h5>
+      <h4>Chapter 17</h4>
+      <button type="button">UPDATE PROGRESS</button>
+    </div>
+  );
+}
+
+export default Progress;
