+++ src/components/Status.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function Status() {
+  return (
+    <div className="status">
+      <div className="circle" />
+      <div className="stats">
+        <p className="percentage">64%</p>
+        <p className="completed">Completed</p>
+      </div>
+    </div>
+  );
+}
+
+export default Status;
